fix(api): live-Demo compat round 2 — error handler, tolerant decimals, volume chart, float-typed counts
Six more bugs caught by a 44-endpoint live smoke test against CoinGecko Demo,
on top of the ISO-8601 / SearchCategoryHit fixes from the first round:

1) No typed error exceptions. /coins/{id} with an unknown id returned
   HttpRequestException from EnsureSuccessStatusCode instead of
   CoinGeckoNotFoundException; all other non-2xx statuses had the same gap.
   Added CoinGeckoErrorHandler as a DelegatingHandler that translates
   4xx/5xx to the appropriate typed exception with status code + body +
   request id. Handler order is now Telemetry → Error → Auth → Plan →
   RateLimit → Retry → Primary, so transient 5xx are still retried before
   becoming exceptions.

2) PoolAttributes.base_token_price_native_currency is returned as "NaN"
   (string) for freshly-created onchain pools with no liquidity. The
   default decimal? parser + AllowReadingFromString threw on those. Added
   TolerantDecimalConverter<decimal?> that maps NaN / null / empty to null
   and registered it globally in CoinGeckoJsonContext.

3) Nft.number_of_unique_addresses is returned as a float (e.g. 173.0) but
   was typed long?. Changed to decimal? (CoinCommunityData.RedditAccountsActive48h
   has the same shape and is also changed).

4) /exchanges/{id}/volume_chart returns [ms_as_number, volume_as_string]
   mixed tuples. The string[][] deserialization we used failed on the int
   at position 0. ExchangesClient now parses these two endpoints via
   JsonDocument and extracts each position by JsonValueKind.

5) Sample used a fixed 2023-12-31 date for history which falls outside the
   Demo-plan 365-day window and returned a business-logic 401. Sample now
   uses a date 30 days ago.

6) AiAgentHub.NftSummary.Holders was long? and is kept that way; the
   projection now casts the decimal? from Nft.NumberOfUniqueAddresses.

Smoke results: 44/44 endpoints pass against live Demo (plan-gating +
error-handling cases included); 204/204 unit tests green on both TFMs.
Sample now exercises 44 endpoints and is suitable as a pre-release smoke
harness (COINGECKO_API_KEY → dotnet run --project samples/...Console).

+++ src/CoinGecko.Api/Models/Coins/CoinCommunityData.cs
@@ -15,8 +15,8 @@ public sealed class CoinCommunityData
     [JsonPropertyName("reddit_average_comments_48h")] public decimal? RedditAverageComments48h { get; init; }
     /// <summary>Reddit subscribers.</summary>
     [JsonPropertyName("reddit_subscribers")] public long? RedditSubscribers { get; init; }
-    /// <summary>Reddit accounts active in the past 48h.</summary>
-    [JsonPropertyName("reddit_accounts_active_48h")] public long? RedditAccountsActive48h { get; init; }
+    /// <summary>Reddit accounts active in the past 48h. (CoinGecko returns this as a float, e.g. <c>0.0</c>.)</summary>
+    [JsonPropertyName("reddit_accounts_active_48h")] public decimal? RedditAccountsActive48h { get; init; }
     /// <summary>Telegram channel user count.</summary>
     [JsonPropertyName("telegram_channel_user_count")] public long? TelegramChannelUserCount { get; init; }
 }
